Data.Struct

Description

Documentation

class Struct t where

Methods

struct :: p t -> Dict (Coercible (t s) (Object s))

data Object s

destruct :: Struct t => t s -> SmallMutableArray# s Any

construct :: Struct t => SmallMutableArray# s Any -> t s

eqStruct :: Struct t => t s -> t s -> Bool

alloc :: (PrimMonad m, Struct t) => Int -> m (t (PrimState m))

Allocate a structure made out of n slots. Initialize the structure before proceeding!

pattern Nil :: () => Struct t => t s

Truly imperative.

isNil :: Struct t => t s -> Bool

data Slot x y

A Slot is a reference to another unboxed mutable object.

class Precomposable t where

We can compose slots to get a nested slot or field accessor

Methods

(#) :: Slot x y -> t y z -> t x z

slot :: Int -> Slot s t

get :: (PrimMonad m, Struct x, Struct y) => Slot x y -> x (PrimState m) -> m (y (PrimState m))

Get the value from a Slot

set :: (PrimMonad m, Struct x, Struct y) => Slot x y -> x (PrimState m) -> y (PrimState m) -> m ()

Set the value of a Slot

data Field x a

A Field is a reference from a struct to a normal Haskell data type.

field :: Int -> Field s a

Store the reference to the Haskell data type in a normal field

unboxedField :: Prim a => Int -> Int -> Field s a

Store the reference in the nth slot of the nth argument, treated as a MutableByteArray

getField :: (PrimMonad m, Struct x) => Field x a -> x (PrimState m) -> m a

Get the value of a field in a struct

setField :: (PrimMonad m, Struct x) => Field x a -> x (PrimState m) -> a -> m ()

Set the value of a field in a struct