Data.Struct.Internal

Description

Documentation

data Dict p where

Constructors

Dict :: p => Dict p

st :: PrimMonad m => ST (PrimState m) a -> m a

Run an ST calculation inside of a PrimMonad. This lets us avoid dispatching everything through the PrimMonad dictionary.

class Struct t where

Methods

struct :: p t -> Dict (Coercible (t s) (Object s))

data Object s

Constructors

Object

Fields

runObject :: SmallMutableArray# s Any

destruct :: Struct t => t s -> SmallMutableArray# s Any

construct :: Struct t => SmallMutableArray# s Any -> t s

unsafeCoerceStruct :: (Struct x, Struct y) => x s -> y s

eqStruct :: Struct t => t s -> t s -> Bool

pattern Struct :: () => Struct t => SmallMutableArray# s (Any *) -> t s

alloc :: (PrimMonad m, Struct t) => Int -> m (t (PrimState m))

Allocate a structure made out of n slots. Initialize the structure before proceeding!

Tony Hoare's billion dollar mistake

nil :: Object s

isNil :: Struct t => t s -> Bool

pattern Nil :: () => Struct t => t s

Truly imperative.

Faking SmallMutableArrayArray#s

writeSmallMutableArraySmallArray# :: SmallMutableArray# s Any -> Int# -> SmallMutableArray# s Any -> State# s -> State# s

readSmallMutableArraySmallArray# :: SmallMutableArray# s Any -> Int# -> State# s -> (#State# s, SmallMutableArray# s Any#)

writeMutableByteArraySmallArray# :: SmallMutableArray# s Any -> Int# -> MutableByteArray# s -> State# s -> State# s

readMutableByteArraySmallArray# :: SmallMutableArray# s Any -> Int# -> State# s -> (#State# s, MutableByteArray# s#)

Field Accessors

data Slot x y

A Slot is a reference to another unboxed mutable object.

Constructors

Slot (forall s. SmallMutableArray# s Any -> State# s -> (#State# s, SmallMutableArray# s Any#)) (forall s. SmallMutableArray# s Any -> SmallMutableArray# s Any -> State# s -> State# s)

class Precomposable t where

We can compose slots to get a nested slot or field accessor

Methods

(#) :: Slot x y -> t y z -> t x z

slot :: Int -> Slot s t

get :: (PrimMonad m, Struct x, Struct y) => Slot x y -> x (PrimState m) -> m (y (PrimState m))

Get the value from a Slot

set :: (PrimMonad m, Struct x, Struct y) => Slot x y -> x (PrimState m) -> y (PrimState m) -> m ()

Set the value of a Slot

data Field x a

A Field is a reference from a struct to a normal Haskell data type.

Constructors

Field (forall s. SmallMutableArray# s Any -> State# s -> (#State# s, a#)) (forall s. SmallMutableArray# s Any -> a -> State# s -> State# s)

field :: Int -> Field s a

Store the reference to the Haskell data type in a normal field

unboxedField :: Prim a => Int -> Int -> Field s a

Store the reference in the nth slot of the nth argument, treated as a MutableByteArray

getField :: (PrimMonad m, Struct x) => Field x a -> x (PrimState m) -> m a

Get the value of a field in a struct

setField :: (PrimMonad m, Struct x) => Field x a -> x (PrimState m) -> a -> m ()

Set the value of a field in a struct