Data.Struct.LinkCut

Description

Documentation

data LinkCut a s

new :: (PrimMonad m, Monoid a) => a -> m (LinkCut a (PrimState m))

O(1). Allocate a new link-cut tree with a given monoidal summary.

link :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> LinkCut a (PrimState m) -> m ()

O(log n). link v w inserts v which must be the root of a tree in as a child of w. v and w must not be connected.

cut :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> m ()

O(log n). cut v removes the linkage between v upwards to whatever tree it was in, making v a root node.

Repeated calls on the same value without intermediate accesses are O(1).

root :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> m (LinkCut a (PrimState m))

O(log n). Find the root of a tree.

Repeated calls on the same value without intermediate accesses are O(1).

cost :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> m a

O(log n). cost v computes the root-to-leaf path cost of v under whatever Monoid was built into the tree.

Repeated calls on the same value without intermediate accesses are O(1).

parent :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> m (LinkCut a (PrimState m))

O(log n). Find the parent of a node.

This will return Nil if the parent does not exist.

Repeated calls on the same value without intermediate accesses are O(1).

connected :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> LinkCut a (PrimState m) -> m Bool

O(log n). connected v w determines if v and w inhabit the same tree.