Data.Struct.Order

Description

Documentation

newtype Order s

This structure maintains an order-maintenance structure as two levels of list-labeling.

The upper labeling scheme holds (n / log w) elements in a universe of size w^2, operating in O(log n) amortized time per operation.

It is accelerated by an indirection structure where each smaller universe has size O(log n), but O(1) expected update cost, so we can charge rebuilds to the upper structure to the lower structure.

This means that inserts are O(1) amortized, while comparisons remain O(1) worst-case.

Constructors

Order

Fields

runOrder :: Object s

newOrder :: PrimMonad m => m (Order (PrimState m))

class Struct t => IsLabel t where

Minimal complete definition

key, next, prev, compareM, delete, insertAfter, cutBefore, cutAfter, least, greatest

Methods

delete :: PrimMonad m => t (PrimState m) -> m ()

O(1). Remove a label

insertAfter :: PrimMonad m => t (PrimState m) -> m (t (PrimState m))

value :: PrimMonad m => Order (PrimState m) -> m (Key, Key)