Data.Struct.Internal.LinkCut

Description

Amortized Link-Cut trees via splay trees based on Tarjan's little book.

Documentation

newtype LinkCut a s

Constructors

LinkCut (Object s)

path :: Slot (LinkCut a) (LinkCut a)

parent :: Slot (LinkCut a) (LinkCut a)

left :: Slot (LinkCut a) (LinkCut a)

right :: Slot (LinkCut a) (LinkCut a)

value :: Field (LinkCut a) a

summary :: Field (LinkCut a) a

new :: (PrimMonad m, Monoid a) => a -> m (LinkCut a (PrimState m))

O(1). Allocate a new link-cut tree with a given monoidal summary.

cut :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> m ()

O(log n). cut v removes the linkage between v upwards to whatever tree it was in, making v a root node.

Repeated calls on the same value without intermediate accesses are O(1).

link :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> LinkCut a (PrimState m) -> m ()

O(log n). link v w inserts v which must be the root of a tree in as a child of w. v and w must not lie in the same tree.

cost :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> m a

O(log n). cost v computes the root-to-leaf path cost of v under whatever Monoid was built into the tree.

Repeated calls on the same value without intermediate accesses are O(1).

root :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> m (LinkCut a (PrimState m))

O(log n). Find the root of a tree.

Repeated calls on the same value without intermediate accesses are O(1).

up :: (PrimMonad m, Monoid a) => LinkCut a (PrimState m) -> m (LinkCut a (PrimState m))

O(log n). Move upward one level.

This will return Nil if the parent is not available.

Note: Repeated calls on the same value without intermediate accesses are O(1).

summarize :: Monoid a => LinkCut a s -> ST s a

O(1)

access :: Monoid a => LinkCut a s -> ST s ()

O(log n)

splay :: Monoid a => LinkCut a s -> ST s ()

O(log n). Splay within an auxiliary tree