Data.Struct.Order

Description

Documentation

data Order s

This structure maintains an order-maintenance structure as two levels of list-labeling.

The upper labeling scheme holds (n / log w) elements in a universe of size w^2, operating in O(log n) amortized time per operation.

It is accelerated by an indirection structure where each smaller universe has size O(log n), but O(1) expected update cost, so we can charge rebuilds to the upper structure to the lower structure.

This means that inserts are O(1) amortized, while comparisons remain O(1) worst-case.

new :: PrimMonad m => m (Order (PrimState m))

insertAfter :: PrimMonad m => Order (PrimState m) -> m (Order (PrimState m))

delete :: PrimMonad m => Order (PrimState m) -> m ()